+++ lib/generators/raw_generator.dart
@@ -31,13 +31,13 @@ const _${flavor == null ? '' : '${flavor}_'}raw_resources = (
     ..sort(sortFilesByName);
   for (final file in files) {
     final relativePath = getRelativePath(
-      file.absolute.uri.path,
+      file.absolute.unixPath,
       getCurrentPath(),
     );
     final fullPath =
         (package == null ? '' : 'packages/$package/') + relativePath;
     buffer.writeln(
-      "  ${safeName(basenameWithoutExtension(file.path))}: '$fullPath',",
+      "  ${safeName(basenameWithoutExtension(file.unixPath))}: '$fullPath',",
     );
   }
   buffer.writeln(');');