+++ lib/generators/vector_generator.dart
@@ -27,19 +27,19 @@ const _${flavor == null ? '' : '${flavor}_'}vector_resources = (
 """);
   final files = directory.listSync().where((element) {
     return element is File &&
-        extension(element.path) == '.svg' &&
+        extension(element.unixPath) == '.svg' &&
         !element.isHidden;
   }).toList()
     ..sort(sortFilesByName);
   for (final file in files) {
     final relativePath = getRelativePath(
-      file.absolute.uri.path,
+      file.absolute.unixPath,
       getCurrentPath(),
     );
     final fullPath =
         (package == null ? '' : 'packages/$package/') + relativePath;
     buffer.writeln(
-      "  ${safeName(basenameWithoutExtension(file.path))}: '$fullPath',",
+      "  ${safeName(basenameWithoutExtension(file.unixPath))}: '$fullPath',",
     );
   }
   buffer.writeln(');');