+++ lib/util/extensions/file_ext.dart
@@ -6,4 +6,8 @@ extension FileSystemEntityExtension on FileSystemEntity {
   bool get isHidden {
     return basename(path).startsWith(RegExp('[._]'));
   }
+
+  String get unixPath {
+    return path.replaceAll(r'\', '/');
+  }
 }