+++ lib/util/sort_algorithm.dart
@@ -1,7 +1,8 @@
 import 'dart:io';
 
 import 'package:path/path.dart';
+import 'package:resources_generator/util/extensions/file_ext.dart';
 
 int sortFilesByName(FileSystemEntity a, FileSystemEntity b) {
-  return basename(a.path).compareTo(basename(b.path));
+  return basename(a.unixPath).compareTo(basename(b.unixPath));
 }